Fixed https://github.com/kit-ipe/WAVE-Client-Renderer/issues/32
- The so-called CORS fixed!?! (https://github.com/kit-ipe/WAVE-Client-Renderer/commit/4f6cc08639f8caf45def3cc740630c872b064b4e) earlier was introducing a redundant
data loading in which I fixed in this commit. The initial data loading
at the volumeraycaster is setup as a image containers, loading a dummy
image. I didn't remove the whole loading section due to the tightly
coupled structure that involves VRC dispatchers---not really relevant
as we should move on to use more stable JQuery event handler etc.

Signed-off-by: Nicholas Tan Jerome <nicholas.jerome@kit.edu>